+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.9
 
-WORKDIR /code
+WORKDIR /python
 
 COPY ./requirements.txt /requirements.txt
