+++ models/models.py
@@ -1,9 +1,9 @@
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, Float, TIMESTAMP
+from sqlalchemy import Column, Integer, String, Float, TIMESTAMP, MetaData
 from datetime import datetime
 
 Base = declarative_base()
-
+metadata = MetaData()
 class CurrencyRates(Base):
     __tablename__ = 'currency_rates'
     id = Column(Integer, primary_key=True)
